requirements.txt
aiogram
aiohttp
dspy
google-genai
python-dotenv
